cart - remove from the cart functionality, added sweet-alert plugin

+++ resources/js/components/Cart.vue
@@ -6,16 +6,20 @@
             </div>
         </div>
         <div class="row">
-            <div class="col-lg-12 cart-item" v-for="item in cart.items" :key="item.id">
+            <div class="col-lg-12 cart-item" v-for="item in cartItems" :key="item.id">
                 <div class="flexbox">
                     <img :src="item.product.image" width="64"/>
                     <div class="prod-name">
                         <h4>{{item.product.name}}</h4>
                         <p>{{item.product.category.name}}</p>
                     </div>
-                    <input type="number" :value="item.qty"/>
+                    <h6>Rs. {{item.product.price}}</h6>
+                    <input type="number" :value="item.qty" disabled/>
                     <button class="btn btn-sm btn-danger" v-on:click="remove(item)">Remove</button>
                 </div>
+            </div>
+            <div class="col-lg-12" v-if="cartItems.length == 0">
+                <div class="alert alert-warning text-center">Your cart is empty. Add items from the product page.</div>
             </div>
         </div>
     </div>
@@ -25,22 +29,29 @@
 export default{
     data(){
         return{
-
+            cartItems : []
         }
     },
     props: [
         'cart'
     ],
     mounted(){
-        console.log(this.cart);
+        this.cartItems = this.cart.items;
     },
     methods: {
         remove : function(item){
-            axios.delete('/cart/remove-item/'+item.id).then((response) => {
-                console.log(response.data);
-            }).catch((error) => {
-                alert('error');
-            });
+            this.ask(null,"This item will be removed from the cart!", function (action){
+                if(action.isConfirmed){
+                    console.log(this.cartItems);
+                    axios.delete('/cart/remove-item/'+item.id).then((response) => {
+                        this.cartItems = this.cartItems.filter(function (ArItem){
+                            return ArItem.id !== item.id;
+                        });
+                        this.showAlert(response.data.message);
+                    }).catch((error) => {});
+                }
+            }.bind(this));
+
         }
     }
 }
